git-svn-id: http://svn.voreen.org/public/voreen-snapshot@5 d84f7722-cf57-0410-9e8f-1be463e99fdd

+++ ext/tgt/tgt_gl.cpp
@@ -33,7 +33,7 @@ GLenum _lGLError(int line, const char* file) {
 
     if (err != GL_NO_ERROR) {
         char title[100];
-        sprintf(title, "OpenGL-Error %d:", err);
+        sprintf(title, "OpenGL-Error %d:", static_cast<int>(err));
         const GLubyte* exp = gluErrorString(err);
 
         std::ostringstream tmp1, tmp2, loggerCat;
